refactor GROMACS log analysis
- put into a new GROMACS specific module
- put code in analyze_run into separate function
